fix: sider plugin
