+++ src/components/organisms/Sider/Sider.tsx
@@ -5,7 +5,6 @@ import {Popover, Tooltip} from 'antd';
 
 import {ReactComponent as ExecutorsIcon} from '@assets/executor.svg';
 import {ReactComponent as SourcesIcon} from '@assets/sources.svg';
-import {ReactComponent as StatusPageIcon} from '@assets/status-page-icon.svg';
 import {ReactComponent as TestSuitesIcon} from '@assets/test-suites-icon.svg';
 import {ReactComponent as Logo} from '@assets/testkube-symbol-color.svg';
 import {ReactComponent as TestsIcon} from '@assets/tests-icon.svg';
@@ -21,6 +20,8 @@ import {FullWidthSpace, Text} from '@custom-antd';
 import {useDashboardNavigate} from '@hooks/useDashboardNavigate';
 
 import {ReactComponent as SettingIcon} from '@icons/setting.svg';
+
+import {usePluginSlotList} from '@plugins/hooks';
 
 import Colors from '@styles/Colors';
 
@@ -42,89 +43,114 @@ const DEFAULT_ICON_STYLE = {
   fontSize: 24,
 };
 
-const getRoutes = (showSocialLinksInSider: boolean, isCloudFeatureEnabled: boolean) => [
+const getRoutes = (showSocialLinksInSider: boolean) => [
   {
-    path: '/tests',
-    icon: TestsIcon,
-    title: 'Tests',
-    transition: {
-      classNames: 'item2',
+    value: {
+      path: '/tests',
+      icon: TestsIcon,
+      title: 'Tests',
+      transition: {
+        classNames: 'item2',
+      },
+    },
+    metadata: {
+      order: 7,
+      visible: () => true,
     },
   },
   {
-    path: '/test-suites',
-    icon: TestSuitesIcon,
-    title: 'Test Suites',
-    transition: {
-      classNames: 'item',
+    value: {
+      path: '/test-suites',
+      icon: TestSuitesIcon,
+      title: 'Test Suites',
+      transition: {
+        classNames: 'item',
+      },
+    },
+    metadata: {
+      order: 6,
+      visible: () => true,
     },
   },
   {
-    path: '/executors',
-    icon: ExecutorsIcon,
-    title: 'Executors',
-    transition: {
-      classNames: 'item',
+    value: {
+      path: '/executors',
+      icon: ExecutorsIcon,
+      title: 'Executors',
+      transition: {
+        classNames: 'item',
+      },
+    },
+    metadata: {
+      order: 5,
+      visible: () => true,
     },
   },
   {
-    path: '/triggers',
-    icon: TriggersIcon,
-    title: 'Triggers',
-    transition: {
-      classNames: 'item',
+    value: {
+      path: '/triggers',
+      icon: TriggersIcon,
+      title: 'Triggers',
+      transition: {
+        classNames: 'item',
+      },
+    },
+    metadata: {
+      order: 4,
+      visible: () => true,
     },
   },
   {
-    path: '/webhooks',
-    icon: WebhooksIcon,
-    title: 'Webhooks',
-    transition: {
-      classNames: 'item',
+    value: {
+      path: '/webhooks',
+      icon: WebhooksIcon,
+      title: 'Webhooks',
+      transition: {
+        classNames: 'item',
+      },
+    },
+    metadata: {
+      order: 3,
+      visible: () => true,
     },
   },
   {
-    path: '/sources',
-    icon: SourcesIcon,
-    title: 'Sources',
-    transition: {
-      classNames: 'item',
+    value: {
+      path: '/sources',
+      icon: SourcesIcon,
+      title: 'Sources',
+      transition: {
+        classNames: 'item',
+      },
+    },
+    metadata: {
+      order: 2,
+      visible: () => true,
     },
   },
-  ...(isCloudFeatureEnabled
-    ? [
-        {
-          path: '/status-page',
-          icon: StatusPageIcon,
-          title: 'Status Page',
-          transition: {
-            classNames: 'item',
-          },
-          active: /status-page/,
-        },
-      ]
-    : []),
 
-  ...(showSocialLinksInSider
-    ? []
-    : [
-        {
-          path: '/settings',
-          icon: SettingIcon,
-          title: 'Settings',
-          transition: {
-            classNames: 'item',
-          },
-          additionalClassName: 'settings-icon',
-          active: /environment-management/,
-        },
-      ]),
+  {
+    value: {
+      path: '/settings',
+      icon: SettingIcon,
+      title: 'Settings',
+      transition: {
+        classNames: 'item',
+      },
+      additionalClassName: 'settings-icon',
+      active: /environment-management/,
+    },
+    metadata: {
+      order: -Infinity,
+      visible: () => !showSocialLinksInSider,
+    },
+  },
 ];
 
 const Sider: React.FC = () => {
-  const {showLogoInSider, showSocialLinksInSider, isCloudFeatureEnabled} = useContext(DashboardContext);
+  const {showLogoInSider, showSocialLinksInSider} = useContext(DashboardContext);
   const openSettings = useDashboardNavigate('/settings');
-
+  const siderMenuList = usePluginSlotList('statusPageSiderMenuItem', getRoutes(showSocialLinksInSider));
   const otherMenuItems = [
     {
       icon: 'cog',
@@ -177,7 +203,7 @@ const Sider: React.FC = () => {
   ];
 
   const renderedMenuItems = useMemo(() => {
-    return getRoutes(showSocialLinksInSider, isCloudFeatureEnabled).map(route => {
+    return siderMenuList.map(route => {
       const {icon: MenuIcon, path, title, active} = route;
 
       return (
@@ -196,7 +222,7 @@ const Sider: React.FC = () => {
         </StyledSiderLink>
       );
     });
-  }, [showSocialLinksInSider, isCloudFeatureEnabled]);
+  }, [showSocialLinksInSider, siderMenuList]);
 
   const renderedOtherMenuItems = useMemo(() => {
     return otherMenuItems.map(otherMenuItem => {
