+++ src/contexts/DashboardContext.ts
@@ -9,7 +9,6 @@ interface DashboardContextData {
   showLogoInSider: boolean;
   showSocialLinksInSider: boolean;
   showTestkubeCloudBanner: boolean;
-  isCloudFeatureEnabled: boolean;
 }
 
 const DashboardContext = createContext<DashboardContextData>(undefined!);
